Move a bunch of types from Spar.API to new Spar.API.Types.
